Fix bug calculation price apply voucher by percent and direct amount

+++ src/main/java/com/swp/backend/model/BookingApplyVoucherModel.java
@@ -6,12 +6,12 @@ import lombok.*;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class BookingApplyVoucherModel extends BookingModel{
-    private int newPrice;
-    private int discountPrice;
+public class BookingApplyVoucherModel extends BookingModel {
+    private float newPrice;
+    private float discountPrice;
 
     @Builder
-    public BookingApplyVoucherModel(int slotId, String refSubYard, int price, String date,  int newPrice, int discountPrice) {
+    public BookingApplyVoucherModel(int slotId, String refSubYard, int price, String date, float newPrice, float discountPrice) {
         super(slotId, refSubYard, price, date);
         this.newPrice = newPrice;
         this.discountPrice = discountPrice;
